assignment 3 complete (role based access)

+++ routes/dishRouter.js
@@ -14,7 +14,6 @@ dishRouter.use(bodyParser.json());
 dishRouter.route('/')
 .get((req,res,next) => {
     Dishes.find({})
-    .populate('comments.author') //dishes will now have author from user document now
     .then((dishes) => {
         res.statusCode = 200;
         res.setHeader('Content-Type', 'application/json');
@@ -23,27 +22,45 @@ dishRouter.route('/')
     .catch((err) => next(err));
 })
 .post(authenticate.verifyUser,  (req, res, next) => {
-    Dishes.create(req.body)
-    .then((dish) => {
-        console.log('Dish Created ', dish);
-        res.statusCode = 200;
-        res.setHeader('Content-Type', 'application/json');
-        res.json(dish);
-    }, (err) => next(err))
-    .catch((err) => next(err));
+    if (authenticate.verifyAdmin(req.user)) {
+        Dishes.create(req.body)
+        .then((dish) => {
+            console.log('Dish Created ', dish);
+            res.statusCode = 200;
+            res.setHeader('Content-Type', 'application/json');
+            res.json(dish);
+        }, (err) => next(err))
+        .catch((err) => next(err));
+    } else {
+        err = new Error('Only Admin users can perform this operation');
+        err.status = 403;
+        return next(err);
+    }
 })
 .put(authenticate.verifyUser, (req, res, next) => {
-    res.statusCode = 403;
-    res.end('PUT operation not supported on /dishes');
+    if (authenticate.verifyAdmin(req.user)) {
+        res.statusCode = 403;
+        res.end('PUT operation not supported on /dishes');
+    } else {
+        err = new Error('Only Admin users can perform this operation');
+        err.status = 403;
+        return next(err);
+    }
 })
 .delete(authenticate.verifyUser, (req, res, next) => {
-    Dishes.remove({})
-    .then((resp) => {
-        res.statusCode = 200;
-        res.setHeader('Content-Type', 'application/json');
-        res.json(resp);
-    }, (err) => next(err))
-    .catch((err) => next(err));    
+    if (authenticate.verifyAdmin(req.user)) {
+        Dishes.remove({})
+        .then((resp) => {
+            res.statusCode = 200;
+            res.setHeader('Content-Type', 'application/json');
+            res.json(resp);
+        }, (err) => next(err))
+        .catch((err) => next(err));  
+    } else {
+        err = new Error('Only Admin users can perform this operation');
+        err.status = 403;
+        return next(err);
+    }
 });
 
 dishRouter.route('/:dishId')
@@ -58,28 +75,46 @@ dishRouter.route('/:dishId')
     .catch((err) => next(err));
 })
 .post(authenticate.verifyUser, (req, res, next) => {
-    res.statusCode = 403;
-    res.end('POST operation not supported on /dishes/'+ req.params.dishId);
+    if (authenticate.verifyAdmin(req.user)) {
+        res.statusCode = 403;
+        res.end('POST operation not supported on /dishes/'+ req.params.dishId);
+    } else {
+        err = new Error('Only Admin users can perform this operation');
+        err.status = 403;
+        return next(err);
+    }
 })
 .put(authenticate.verifyUser, (req, res, next) => {
-    Dishes.findByIdAndUpdate(req.params.dishId, {
-        $set: req.body
-    }, { new: true })
-    .then((dish) => {
-        res.statusCode = 200;
-        res.setHeader('Content-Type', 'application/json');
-        res.json(dish);
-    }, (err) => next(err))
-    .catch((err) => next(err));
+    if (authenticate.verifyAdmin(req.user)) {
+        Dishes.findByIdAndUpdate(req.params.dishId, {
+            $set: req.body
+        }, { new: true })
+        .then((dish) => {
+            res.statusCode = 200;
+            res.setHeader('Content-Type', 'application/json');
+            res.json(dish);
+        }, (err) => next(err))
+        .catch((err) => next(err));
+    } else {
+        err = new Error('Only Admin users can perform this operation');
+        err.status = 403;
+        return next(err);
+    }
 })
 .delete(authenticate.verifyUser, (req, res, next) => {
-    Dishes.findByIdAndRemove(req.params.dishId)
-    .then((resp) => {
-        res.statusCode = 200;
-        res.setHeader('Content-Type', 'application/json');
-        res.json(resp);
-    }, (err) => next(err))
-    .catch((err) => next(err));
+    if (authenticate.verifyAdmin(req.user)) {
+        Dishes.findByIdAndRemove(req.params.dishId)
+        .then((resp) => {
+            res.statusCode = 200;
+            res.setHeader('Content-Type', 'application/json');
+            res.json(resp);
+        }, (err) => next(err))
+        .catch((err) => next(err));
+    } else {
+        err = new Error('Only Admin users can perform this operation');
+        err.status = 403;
+        return next(err);
+    }
 });
 
 dishRouter.route('/:dishId/comments')
@@ -134,26 +169,32 @@ dishRouter.route('/:dishId/comments')
         + req.params.dishId + '/comments');
 })
 .delete(authenticate.verifyUser, (req, res, next) => {
-    Dishes.findById(req.params.dishId)
-    .then((dish) => {
-        if (dish != null) {
-            for (var i = (dish.comments.length -1); i >= 0; i--) {
-                dish.comments.id(dish.comments[i]._id).remove();
+    if (authenticate.verifyAdmin(req.user)) {
+        Dishes.findById(req.params.dishId)
+        .then((dish) => {
+            if (dish != null) {
+                for (var i = (dish.comments.length -1); i >= 0; i--) {
+                    dish.comments.id(dish.comments[i]._id).remove();
+                }
+                dish.save()
+                .then((dish) => {
+                    res.statusCode = 200;
+                    res.setHeader('Content-Type', 'application/json');
+                    res.json(dish);                
+                }, (err) => next(err));
             }
-            dish.save()
-            .then((dish) => {
-                res.statusCode = 200;
-                res.setHeader('Content-Type', 'application/json');
-                res.json(dish);                
-            }, (err) => next(err));
-        }
-        else {
-            err = new Error('Dish ' + req.params.dishId + ' not found');
-            err.status = 404;
-            return next(err);
-        }
-    }, (err) => next(err))
-    .catch((err) => next(err));    
+            else {
+                err = new Error('Dish ' + req.params.dishId + ' not found');
+                err.status = 404;
+                return next(err);
+            }
+        }, (err) => next(err))
+        .catch((err) => next(err));  
+    } else {
+        err = new Error('Only Admin users can perform this operation');
+        err.status = 403;
+        return next(err);
+    }      
 });
 
 dishRouter.route('/:dishId/comments/:commentId')
@@ -187,33 +228,39 @@ dishRouter.route('/:dishId/comments/:commentId')
 .put(authenticate.verifyUser, (req, res, next) => {
     Dishes.findById(req.params.dishId)
     .then((dish) => {
-        if (dish != null && dish.comments.id(req.params.commentId) != null) {
-            if (req.body.rating) {
-                dish.comments.id(req.params.commentId).rating = req.body.rating;
+        if (JSON.stringify(dish.comments.id(req.params.commentId).author) === JSON.stringify(req.user._id)) {
+            if (dish != null && dish.comments.id(req.params.commentId) != null) {
+                if (req.body.rating) {
+                    dish.comments.id(req.params.commentId).rating = req.body.rating;
+                }
+                if (req.body.comment) {
+                    dish.comments.id(req.params.commentId).comment = req.body.comment;                
+                }
+                dish.save()
+                .then((dish) => {
+                    Dishes.findById(dish._id)
+                        .populate('comments.author')
+                        .then((dish) => {
+                            res.statusCode = 200;
+                            res.setHeader('Content-Type', 'application/json');
+                            res.json(dish);                
+                        })
+                }, (err) => next(err));
             }
-            if (req.body.comment) {
-                dish.comments.id(req.params.commentId).comment = req.body.comment;                
+            else if (dish == null) {
+                err = new Error('Dish ' + req.params.dishId + ' not found');
+                err.status = 404;
+                return next(err);
             }
-            dish.save()
-            .then((dish) => {
-                Dishes.findById(dish._id)
-                    .populate('comments.author')
-                    .then((dish) => {
-                        res.statusCode = 200;
-                        res.setHeader('Content-Type', 'application/json');
-                        res.json(dish);                
-                    })
-            }, (err) => next(err));
-        }
-        else if (dish == null) {
-            err = new Error('Dish ' + req.params.dishId + ' not found');
-            err.status = 404;
+            else {
+                err = new Error('Comment ' + req.params.commentId + ' not found');
+                err.status = 404;
+                return next(err);            
+            }
+        } else {
+            err = new Error('Only author of this comment can perform this operation ');
+            err.status = 403;
             return next(err);
-        }
-        else {
-            err = new Error('Comment ' + req.params.commentId + ' not found');
-            err.status = 404;
-            return next(err);            
         }
     }, (err) => next(err))
     .catch((err) => next(err));
@@ -221,28 +268,34 @@ dishRouter.route('/:dishId/comments/:commentId')
 .delete(authenticate.verifyUser,     (req, res, next) => {
     Dishes.findById(req.params.dishId)
     .then((dish) => {
-        if (dish != null && dish.comments.id(req.params.commentId) != null) {
-            dish.comments.id(req.params.commentId).remove();
-            dish.save()
-            .then((dish) => {
-                Dishes.findById(dish._id)
-                    .populate('comments.author')
-                    .then((dish) => {
-                        res.statusCode = 200;
-                        res.setHeader('Content-Type', 'application/json');
-                        res.json(dish);                
-                    })
-            }, (err) => next(err));
-        }
-        else if (dish == null) {
-            err = new Error('Dish ' + req.params.dishId + ' not found');
-            err.status = 404;
+        if (JSON.stringify(dish.comments.id(req.params.commentId).author) === JSON.stringify(req.user._id)) { //dish.comments.id(req.params.commentId).author & req.user._id were of object types
+            if (dish != null && dish.comments.id(req.params.commentId) != null) {
+                dish.comments.id(req.params.commentId).remove();
+                dish.save()
+                .then((dish) => {
+                    Dishes.findById(dish._id)
+                        .populate('comments.author')
+                        .then((dish) => {
+                            res.statusCode = 200;
+                            res.setHeader('Content-Type', 'application/json');
+                            res.json(dish);                
+                        })
+                }, (err) => next(err));
+            }
+            else if (dish == null) {
+                err = new Error('Dish ' + req.params.dishId + ' not found');
+                err.status = 404;
+                return next(err);
+            }
+            else {
+                err = new Error('Comment ' + req.params.commentId + ' not found');
+                err.status = 404;
+                return next(err);            
+            }
+        } else {
+            err = new Error('Only author of this comment can perform this operation ');
+            err.status = 403;
             return next(err);
-        }
-        else {
-            err = new Error('Comment ' + req.params.commentId + ' not found');
-            err.status = 404;
-            return next(err);            
         }
     }, (err) => next(err))
     .catch((err) => next(err));
